finished the grouping and filtering and changed the stave class
updated the stave group class to contain information about the previous and the next stave group
currently the stave group in the aspect of checking containment is bullcrap but it want be hard to figure out how to do it
the current treshold to the gap between two stave groups is incorect. it doesnt take to account the fact that the number of lines and the size of the gaps between stave lines so this must be changed quickly
consider using some variation of average of the diff_stave or the number with the most instances.